+++ Forge/changelog.md
@@ -3,3 +3,5 @@
 * Rename toasts with honey and jam
 * Tweaked recipes that used toast to use toast slice instead
 * Tweaked Jar recipe
+* Made Dough recipes shapeless
+* Added Tomato Dough